Mistake of uint8

+++ main_drone.py
@@ -96,7 +96,7 @@ def main() -> int:
 
     dBuf_img0 = np.memmap(os.path.join(ddir, "spectr.mmmp.npy"), mode="w+", shape=(cli_batch_num_save, 480, 200), dtype=np.uint8)
     dBuf_img1 = np.memmap(os.path.join(ddir, "webcam.mmmp.npy"), mode="w+", shape=(cli_batch_num_save, 480, 640), dtype=np.uint8)
-    dBuf_ori  = np.memmap(os.path.join(ddir, "orient.mmmp.npy"), mode="w+", shape=(cli_batch_num_save, 7))
+    dBuf_ori  = np.memmap(os.path.join(ddir, "orient.mmmp.npy"), mode="w+", shape=(cli_batch_num_save, 7), dtype=np.float64)
     logging.info(f"Creating the MemMap files")
 
     #dBuf_img0[:] = 0
